endpoint: introduce hid_property
Introduces a HIDProperty class. A HID property represents a data type
from the HID report descriptor (eg. X, Y, etc axis, buttons, etc).

Each HIDProperty must implement a populate() method which receives a
high-level action and a HID event array. It will populate the HID event
array based on the high-level action.

Example:
    (definition)
    ```python
    class ExampleButtonProperty(HIDProperty):
        ...
        def populate(action, packets):
            for packet in packets:
                if 'buttons' in action:
                    for key, val in action['buttons'].items():
                        setattr(packet, f'b{key}', 1 if val else 0)
    ```

    (usage)
    ```python
    ...

    action = {
        ...
        'buttons': {
            0: True,
            1: False,
            2: True,
        }
        ...
    }

    packets = []
    for _ in range(packet_count):
        packets.append(EventData())

    '''
    endpoint.populate_hid_data() will call populate() for each HID property
    in the endpoint, includes our populate() method defined above
    '''
    endpoint.populate_hid_data(action, packets)

    for packet in packets:
        endpoint.send(endpoint.create_report(packets))
    ```

Note: This is a simple example using buttons, there's not much to do
here, we just set the button value in the events. A example where more
things need to happen is in the axis. I didn't provide the axis as an
example because the code is a bit more involved and harder to read.
This should provide an overview on how the different components are
supposed to be used togather.

The idea is to define a HIDProperty for each HID usage and teach it how
to handle that data.

This patch just introduces the class and makes Endpoint to hold a list
the of HID properties. Integration with the rest of the codebase comes
in the next patches.

Signed-off-by: Filipe Laíns <lains@archlinux.org>