device: use HIDProperty to do the high-level action to HID data conversion
In a previous patch we introduced HIDProperty, classes which subclass
it (eg. AxisProperty) should know how to handle specific HID usages.

This patch replaces the current _simulate_action_* routines with HID
properties. Instead of teaching Device how to handle every single HID
usage we define HID properties.

Each endpoint has a list of properties which is dynamically discovered
from its report descriptor. When it's time to perform an action we
generate the a HID event list (the lenght is based on the action
duration and the device report rate) and ask each endpoint to populate.

Signed-off-by: Filipe Laíns <lains@archlinux.org>